refactor: move INestedStyleValue to the central form models
INestedStyleValue lived in a migration file, which was the wrong home for a
type that describes a component's style model. It now sits in
providers/form/models.ts next to IStyleValue, which it builds on.

Drops the styles-to-nested-set migration and its call from the checkbox group
migrator.
