+++ shesha-reactjs/src/providers/form/models.ts
@@ -140,6 +140,16 @@ export interface IStyleValue {
   autoHeight?: boolean | undefined;
 }
 
+/**
+ * The style model of a component that exposes two independent sets of Appearance panels: the
+ * bare-named properties style the component's wrapper, and a nested set styles a repeated child.
+ *
+ * Used by the option components, where the wrapper is the group container and the nested set
+ * styles each option — `INestedStyleValue<'radio'>` gives `radio.border`, `radio.background`, …
+ * alongside the wrapper's own `border`, `background`, ….
+ */
+export type INestedStyleValue<TNested extends string> = IStyleValue & { [K in TNested]?: IStyleValue | undefined };
+
 export interface IInputStyles extends IStyleValue {
   borderSize?: string | number | undefined;
   borderRadius?: string | number | undefined;
